+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Hm version history
 
+## 0.0.3 - Mar 8, 2018
+
+* `Hm#partition(list, of, keys)` to receive two hashes: with matching, and not matching keys;
+* `Hm#transform_values` now work without any keys/pathes passed (just transforming all "terminal" values).
+
 ## 0.0.2 - Feb 24, 2018
 
 * Some profiling and small optimizations;